Fix configuration and client for tests
