+++ src/account/__init__.py
@@ -1,7 +1,1 @@
-from .api import AccountAPI
-from .balance import AccountBalance
-
-__all__ = [
-    'AccountAPI',
-    'AccountBalance'
-]
+from kis.account import *
