+++ src/core/config.py
@@ -1,0 +1,1 @@
+from kis.core.config import *
