+++ src/kis/core/__init__.py
@@ -1,0 +1,25 @@
+"""
+KIS Core 모듈
+
+이 패키지는 한국투자증권 API의 핵심 기능을 제공합니다:
+- API 클라이언트 (KISClient)
+- 인증 관리 (auth)
+- 공통 유틸리티
+
+사용 예시:
+    from kis.core import KISClient
+    client = KISClient()
+"""
+
+from .agent import KIS_Agent
+from .client import KISClient, API_ENDPOINTS
+from .auth import auth
+from .config import KISConfig
+
+__all__ = [
+    'KIS_Agent',
+    'KISClient',
+    'API_ENDPOINTS',
+    'auth',
+    'KISConfig'
+]
