+++ src/kis/program/__init__.py
@@ -1,0 +1,10 @@
+from .api import ProgramTradeAPI
+
+# 기존 호환성을 위한 별칭
+ProgramAPI = ProgramTradeAPI
+ProgramTrade = ProgramTradeAPI
+
+__all__ = [
+    'ProgramAPI',
+    'ProgramTrade'
+]
