+++ src/program/__init__.py
@@ -1,7 +1,1 @@
-from .api import ProgramAPI
-from .trade import ProgramTrade
-
-__all__ = [
-    'ProgramAPI',
-    'ProgramTrade'
-]
+from kis.program import *
