+++ src/strategy/__init__.py
@@ -1,0 +1,1 @@
+from kis.strategy import *
