+++ src/strategy/trigger.py
@@ -1,0 +1,1 @@
+from kis.strategy.trigger import *
