+++ tests/integration/test_account.py
@@ -19,6 +19,10 @@
 import unittest
 import os
 from dotenv import load_dotenv
+import pytest
+
+if not os.getenv('RUN_LIVE_TESTS'):
+    pytest.skip('실제 API 테스트 건너뜀', allow_module_level=True)
 
 from account.api import AccountAPI
 from core.client import KISClient
